cpanel/backend/requirements.txt
Flask~=1.1.2
python-dotenv~=0.15.0
pandas~=1.2.3
flask-cors~=3.0.10